Fixed the Slider/Scrolling Problem
found a way to allow touch devices to scroll the page when touching
the pixijs area by using:

    this.app.renderer.plugins.interaction.autoPreventDefault = false;
    this.app.renderer.view.style['touch-action'] = 'auto';

but this previously had the problem of making it difficult to use
the sliders, because the webpage would start scrolling.  Apparently,
pixi.js allows you to access the original TouchEvent which has the
ability to prevent this scrolling behavior.  This is done in the
interaction handler:

    function onDataPointDragStart(event) {
        event.data.originalEvent.preventDefault();
        /* ... */
    }

So, now you have the best of both worlds!  You can scroll whenever
you are touching the background, and you can STOP scrolling when you
are trying to use one of the sliders!